@@ -1,18 +1,24 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, inject } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { SignInRequest } from '../models/sign-in-request';
 import * as AccountActions from '../stores/account.actions';
 import * as AccountSelectors from '../stores/account.selector';
 import { AccountState } from '../stores/account.state';
-import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
+import { UntypedFormBuilder, UntypedFormGroup, Validators, ReactiveFormsModule } from '@angular/forms';
+import { FormErrorComponent } from '../../shared/form-error/form-error.component';
+import { NgClass, AsyncPipe } from '@angular/common';
+import { RouterLink } from '@angular/router';
 
 @Component({
   selector: 'app-sign-in',
   templateUrl: './sign-in.component.html',
   styleUrls: ['./sign-in.component.css'],
+  imports: [ReactiveFormsModule, FormErrorComponent, NgClass, RouterLink, AsyncPipe],
 })
 export class SignInComponent implements OnInit {
-  constructor(private accountStore: Store<AccountState>, private formBuilder: UntypedFormBuilder) {}
+  private readonly accountStore = inject<Store<AccountState>>(Store);
+  private readonly formBuilder = inject(UntypedFormBuilder);
+
   public signInForm: UntypedFormGroup;
   public isLoading$ = this.accountStore.select(AccountSelectors.isLoading);
 
