@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, inject } from '@angular/core';
 import { Router } from '@angular/router';
 
 @Component({
@@ -7,7 +7,7 @@ import { Router } from '@angular/router';
   styleUrls: ['./unauthorized.component.css'],
 })
 export class UnauthorizedComponent implements OnInit {
-  constructor(private router: Router) {}
+  private readonly router = inject(Router);
 
   ngOnInit(): void {}
 
