@@ -1,21 +1,17 @@
-import { HttpClient } from "@angular/common/http";
-import { environment } from "src/environments/environment";
-import { DashboardData } from "../models/dashboard-data";
-import { Observable } from "rxjs";
-import { Injectable } from "@angular/core";
+import { HttpClient } from '@angular/common/http';
+import { environment } from 'src/environments/environment';
+import { DashboardData } from '../models/dashboard-data';
+import { Observable } from 'rxjs';
+import { Injectable, inject } from '@angular/core';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class DashboardService {
-
-  private dashboardUrl = `${environment.dietPlannerApiUri}/api/dashboard`;
-
-  public constructor(private httpClient: HttpClient) {
-
-  }
+  private readonly httpClient = inject(HttpClient);
+  private readonly dashboardUrl = `${environment.dietPlannerApiUri}/api/dashboard`;
 
   public getDashboardData(): Observable<DashboardData> {
-    return this.httpClient.get<DashboardData>(this.dashboardUrl, { withCredentials: true })
+    return this.httpClient.get<DashboardData>(this.dashboardUrl, { withCredentials: true });
   }
 }
