@@ -1,15 +1,16 @@
-import { Component, EventEmitter, OnInit, Output } from '@angular/core';
-import { NgbDate } from '@ng-bootstrap/ng-bootstrap';
+import { Component, OnInit, output } from '@angular/core';
+import { NgbDate, NgbInputDatepicker } from '@ng-bootstrap/ng-bootstrap';
 import { DatePickerSelection } from '../../models/date-picker-selection';
+import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 
 @Component({
-  selector: 'app-date-selection',
-  templateUrl: './date-selection.component.html',
-  styleUrl: './date-selection.component.scss',
+    selector: 'app-date-selection',
+    templateUrl: './date-selection.component.html',
+    styleUrl: './date-selection.component.scss',
+    imports: [NgbInputDatepicker, ReactiveFormsModule, FormsModule]
 })
 export class DateSelectionComponent implements OnInit {
-  @Output()
-  public selectedDate = new EventEmitter<Date>();
+  public readonly selectedDate = output<Date>();
 
   public dateModel: DatePickerSelection;
 
