@@ -1,5 +1,5 @@
-import { CommonModule, DecimalPipe, NgIf } from '@angular/common';
-import { Component, inject, Input, OnInit, Output } from '@angular/core';
+import { CommonModule, DecimalPipe } from '@angular/common';
+import { Component, inject, OnInit, Output, input } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { Store } from '@ngrx/store';
 import { Dish } from 'src/app/dishes/models/dish';
@@ -14,21 +14,20 @@ import { ItemType } from 'src/app/shared/models/base-item';
 @Component({
   selector: '[app-meal-dish-row-details]',
   templateUrl: 'meal-dish-row-details.component.html',
-  standalone: true,
   imports: [DecimalPipe, CommonModule, FormsModule],
 })
 export class MealDishRowDetailsComponent implements OnInit {
-  @Input() public dish: Dish;
-  @Input() public selectedDate: Date;
-  @Input() public dishProduct: DishProduct;
-  @Input() public mealType: MealType;
-  @Input() public itemIndex: number;
+  public readonly dish = input<Dish>(undefined);
+  public readonly selectedDate = input<Date>(undefined);
+  public readonly dishProduct = input<DishProduct>(undefined);
+  public readonly mealType = input<MealType>(undefined);
+  public readonly itemIndex = input<number>(undefined);
 
   private mealCalendarState: Store<MealCalendarState> = inject(Store<MealCalendarState>);
   public portion: number = 0;
 
   public ngOnInit(): void {
-    this.portion = MealCalendarCalculator.calculatePortion(this.dishProduct);
+    this.portion = MealCalendarCalculator.calculatePortion(this.dishProduct());
   }
 
   public onPortionValueChange(customizedPoritonSize: number): void {
@@ -36,10 +35,10 @@ export class MealDishRowDetailsComponent implements OnInit {
       MealCalendarActions.updateMealItemPortionRequest({
         request: {
           itemType: ItemType.Dish,
-          dishProductId: this.dishProduct.dishProductId,
-          itemProductId: this.dish.mealItemId,
+          dishProductId: this.dishProduct().dishProductId,
+          itemProductId: this.dish().mealItemId,
           customizedPortionMultiplier: customizedPoritonSize / MealCalendarCalculator.defaultPortionSize,
-          date: this.selectedDate,
+          date: this.selectedDate(),
         },
       }),
     );
