@@ -1,4 +1,4 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input, OnInit, inject, input } from '@angular/core';
 import { UntilDestroy } from '@ngneat/until-destroy';
 import { BehaviorSubject } from 'rxjs';
 import * as MealCalendarActions from '../../stores/meals-calendar.actions';
@@ -12,7 +12,6 @@ import { ProductsState } from '../../../products/stores/products.state';
 import { BaseItem } from 'src/app/shared/models/base-item';
 import { MealItemRequest, Meal } from '../../models/meal';
 import { MealItemRowComponent } from './meal-item-row/meal-item-row.component';
-import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { SearchInputComponent } from '../search-input/search-input.component';
 import { MealSummaryRowComponent } from './meal-summary-row/meal-summary-row.component';
@@ -24,10 +23,8 @@ import { MealDishRowDetailsTitleComponent } from './meal-dish-row-details-title/
   selector: 'app-meal-section',
   templateUrl: './meal-section.component.html',
   styleUrls: ['./meal-section.component.css'],
-  standalone: true,
   imports: [
     MealItemRowComponent,
-    CommonModule,
     FormsModule,
     SearchInputComponent,
     MealSummaryRowComponent,
@@ -36,16 +33,14 @@ import { MealDishRowDetailsTitleComponent } from './meal-dish-row-details-title/
   ],
 })
 export class MealSectionComponent implements OnInit {
+  private readonly mealCalendarStore = inject<Store<MealCalendarState>>(Store);
+  private readonly dishStore = inject<Store<DishState>>(Store);
+  private readonly productStore = inject<Store<ProductsState>>(Store);
+
   @Input() public meal: Meal;
-  @Input() calendarDate: Date;
+  readonly calendarDate = input<Date>(undefined);
 
   public searchItem: string;
-
-  constructor(
-    private mealCalendarStore: Store<MealCalendarState>,
-    private dishStore: Store<DishState>,
-    private productStore: Store<ProductsState>,
-  ) {}
 
   public ngOnInit(): void {
     this.dishStore.dispatch(DishActions.loadDishesRequest());
@@ -55,7 +50,7 @@ export class MealSectionComponent implements OnInit {
   public itemAddedToSearchBar(item: BaseItem, mealTypeId: MealType) {
     const addMealRequest: MealItemRequest = {
       mealType: mealTypeId,
-      date: this.calendarDate,
+      date: this.calendarDate(),
       mealItemId: item.mealItemId,
       itemType: item.itemType,
       itemId: item.id,
