@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, inject } from '@angular/core';
 import * as AccountActions from '../stores/account.actions';
 import { AccountState } from '../stores/account.state';
 import { Store } from '@ngrx/store';
@@ -7,9 +7,10 @@ import { Store } from '@ngrx/store';
   selector: 'app-sign-out',
   templateUrl: './sign-out.component.html',
   styleUrls: ['./sign-out.component.css'],
+  standalone: false,
 })
 export class SignOutComponent implements OnInit {
-  constructor(private accountStore: Store<AccountState>) {}
+  private readonly accountStore = inject<Store<AccountState>>(Store);
 
   public ngOnInit(): void {
     this.accountStore.dispatch(AccountActions.signOutRequest());
