@@ -1,5 +1,5 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 import { environment } from 'src/environments/environment';
 import { UserProfile } from '../models/user-profile';
 import { Observable } from 'rxjs';
@@ -9,6 +9,8 @@ import { UserAvatar } from '../models/user-avatar';
   providedIn: 'root',
 })
 export class UserProfileService {
+  private readonly httpClient = inject(HttpClient);
+
   private userProfileApiUrl = `${environment.dietPlannerApiUri}/api/userprofile`;
 
   httpOptions = {
@@ -17,8 +19,6 @@ export class UserProfileService {
       'Content-Type': 'application/json',
     }),
   };
-
-  constructor(private httpClient: HttpClient) { }
 
   public getUserProfile(): Observable<UserProfile> {
     return this.httpClient.get<UserProfile>(this.userProfileApiUrl, { withCredentials: true });
@@ -29,6 +29,10 @@ export class UserProfileService {
   }
 
   public updateUserAvatar(content: string): Observable<UserProfile> {
-    return this.httpClient.patch<UserProfile>(`${this.userProfileApiUrl}/avatar`, { base64Image: content }, this.httpOptions);
+    return this.httpClient.patch<UserProfile>(
+      `${this.userProfileApiUrl}/avatar`,
+      { base64Image: content },
+      this.httpOptions,
+    );
   }
 }
