@@ -1,12 +1,13 @@
-import { Injectable } from '@angular/core';
+import { Injectable, inject } from '@angular/core';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { catchError, exhaustMap, of, switchMap, tap } from 'rxjs';
 import * as DashboardActions from './dashboard.actions';
-import { Router } from '@angular/router';
 import { DashboardService } from '../services/dashboard.service';
 
 @Injectable()
 export class DashboardEffects {
+  private readonly actions$ = inject(Actions);
+  private readonly dashboardService = inject(DashboardService);
 
   getDashboardDataEffect$ = createEffect(() =>
     this.actions$.pipe(
@@ -16,19 +17,16 @@ export class DashboardEffects {
           switchMap(data => {
             return of(DashboardActions.getDashboardDataSuccess({ data }));
           }),
-          catchError(error => of(DashboardActions.getDashboardDataFailed({
-            errorCode: error.status,
-            errorMessage: error.error.message
-          }))),
+          catchError(error =>
+            of(
+              DashboardActions.getDashboardDataFailed({
+                errorCode: error.status,
+                errorMessage: error.error.message,
+              }),
+            ),
+          ),
         );
       }),
     ),
   );
-
-
-  constructor(
-    private actions$: Actions,
-    private router: Router,
-    private dashboardService: DashboardService,
-  ) { }
 }
