@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, inject } from '@angular/core';
 import { Dish } from '../models/dish';
 import { DishService } from '../services/dish.service';
 import { map, take } from 'rxjs';
@@ -7,18 +7,24 @@ import { DishState } from '../stores/dish.state';
 import { Store } from '@ngrx/store';
 import * as DishActions from '../stores/dish.actions';
 import * as DishSelectors from '../stores/dish.selectors';
+import { DishTemplateComponent } from '../dish-template/dish-template.component';
 
 @Component({
   selector: 'app-dish-edit',
   templateUrl: './dish-edit.component.html',
   styleUrls: ['./dish-edit.component.css'],
+  imports: [DishTemplateComponent],
 })
 export class DishEditComponent {
+  private readonly dishStore = inject<Store<DishState>>(Store);
+  private readonly dishService = inject(DishService);
+  private readonly router = inject(ActivatedRoute);
+
   public dish: Dish;
   public submitFunction: (store: any) => void;
   public currentDishes$ = this.dishStore.select(DishSelectors.getDishes);
 
-  constructor(private dishStore: Store<DishState>, private dishService: DishService, private router: ActivatedRoute) {
+  constructor() {
     this.submitFunction = this.getSubmitFunction();
     this.router.params.pipe(take(1)).subscribe(params => {
       this.dishService
