@@ -1,9 +1,8 @@
 import { Component } from '@angular/core';
 
 @Component({
-  selector: '[app-meal-dish-row-details-title]',
-  templateUrl: 'meal-dish-row-details-title.component.html',
-  standalone: true,
-  imports: [],
+    selector: '[app-meal-dish-row-details-title]',
+    templateUrl: 'meal-dish-row-details-title.component.html',
+    imports: []
 })
 export class MealDishRowDetailsTitleComponent {}
